Remove extra ProposalMethods module
This module was used as a 'common' include for ProposalObject
and Proposal switch. It is no longer needed and the methods were moved
to the Proposal model.
